requirements.txt
Flask  
python-dotenv  
azure-ai-formrecognizer  
azure-cognitiveservices-speech  
openai  
ffmpeg-python  
